[Misc] Simplify PoolerOutput and move to `v1/outputs` (#25629)
